load-asset-price-yahoo/requirements.txt
functions-framework==3.*
db-dtypes
pandas
numpy
google-cloud-bigquery
ta
yfinance
